5. Конвертер валюты[#250302]

+++ src/converter/Converter.java
@@ -6,14 +6,14 @@ public class Converter {
         return rsl;
     }
 
-    public static int rubleToDollar(int value) {
-        int rsl = value /60;
+    public static float rubleToDollar(float value) {
+        float rsl = value /60;
         return rsl;
     }
 
     public static void main(String[] args) {
         int euro = Converter.rubleToEuro(140);
-        int dollar = Converter.rubleToDollar(140);
+        float dollar = Converter.rubleToDollar(140);
         System.out.println("140 rubles are " + euro + " euro or " + dollar + " dollar.");
     }
 }
